backend/pyproject.toml
[tool.poetry]
name = "apple_dl"
version = "0.1.0"
description = ""
authors = ["faith <jamienibtong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.35.0"
gamdl = "^2.5"
quart = "^0.20.0"
quart-cors = "^0.8.0"
python-socketio = "^5.13.0"
uvicorn-worker = "^0.3.0"
gunicorn = "^23.0.0"
aiohttp = "^3.12.15"
discord-py = {extras = ["voice"], version = "^2.6.2"}
quart-schema = "^0.22.0"
msgspec = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "apple_dl:run"

[tool.pyright]
venvPath = "."
venv = ".venv"
